% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul_by_three.R
\name{mul_by_three}
\alias{mul_by_three}
\title{Multiply a number by three}
\usage{
mul_by_three(num)
}
\arguments{
\item{num}{A number that needs to be multiplied by 3}
}
\value{
A multiple
}
\description{
Multiply a number by three
}
\examples{
mul_by_three(3)
}
