% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul_by_two.R
\name{mul_by_two}
\alias{mul_by_two}
\title{Multiply a number by 2}
\usage{
mul_by_two(num)
}
\arguments{
\item{num}{a numerical value that
needs to be multiplied by 2}
}
\value{
The value multiplied by 2
}
\description{
Multiply a number by 2
}
\examples{
mul_by_two(2)

}
